requirements.txt
pydantic==1.10.7
PyYAML~=6.0
pytest~=7.3.1